Update `Staker` and `StakingCondition` structs
* Update `Staker` struct
* Update `StakingCondition` struct
* Other minor changes to allow for updates

+++ contracts/base/Staking20Base.sol
@@ -51,7 +51,7 @@ contract Staking20Base is ContractMetadata, Multicall, Ownable, Staking20 {
     uint256 private rewardTokenBalance;
 
     constructor(
-        uint256 _timeUnit,
+        uint80 _timeUnit,
         uint256 _rewardRatioNumerator,
         uint256 _rewardRatioDenominator,
         address _stakingToken,
